Previous approach to the centering the patterns was bottom up and getting messy. Now trying a top down approach which seems to be going a lot better. Lots of code is currenty commented out as it has been broken.
Had to use an intermediary type family called GridPattern' to bring the
dimentionality of the pattern into the type system. This allows us to
better deal with the different index types required when converting into
the intermediary representation.

The plan is to use arrays for the intermediary then write some functions
that deal with the insertion of columns and centering of cursor. Then it
will be trivial to convert this into the tuple representation.